Split main class.
