+++ tests/lint/spelling.ignore-words.txt
@@ -1,0 +1,2 @@
+eventtypes
+
